added all obv3 examples as json_files for testing, finally i did not pursue testing the builders. This will be discussed in the next meeting
