+++ src/tests/json_examples/alignment2.json
@@ -1,0 +1,36 @@
+{
+    "@context": [
+      "https://www.w3.org/2018/credentials/v1",
+      "https://purl.imsglobal.org/spec/ob/v3p0/context-3.0.2.json"
+    ],
+    "id": "http://example.edu/credentials/3732",
+    "type": ["VerifiableCredential", "OpenBadgeCredential"],
+    "issuer": {
+      "id": "https://example.edu/issuers/565049",
+      "type": "Profile",
+      "name": "Example University"
+    },
+    "issuanceDate": "2010-01-01T00:00:00Z",
+    "name": "Example University Degree",
+    "credentialSubject": {
+      "id": "did:example:ebfeb1f712ebc6f1c276e12ec21",
+      "type": "AchievementSubject",
+      "achievement": {
+        "id": "https://1edtech.edu/achievements/1",
+        "type": "Achievement",
+        "criteria": {
+          "narrative": "Cite strong and thorough textual evidence to support analysis of what the text says explicitly as well as inferences drawn from the text, including determining where the text leaves matters uncertain"
+        },
+        "description": "Analyze a sample text",
+        "name": "Text analysis",
+        "alignment": [{
+          "type": "Alignment",
+          "targetCode": "ce-cf4dee18-7cea-443a-b920-158a0762c6bf",
+          "targetFramework": "Edmonds College Course Catalog",
+          "targetName": "Requirements Analysis",
+          "targetType": "ceterms:Credential",
+          "targetUrl": "https://credentialfinder.org/credential/20229/Requirements_Analysis"
+        }]
+      }
+    }
+}
